scripts/python/requirements.txt
Flask>=2.2
google-api-python-client>=2.127.0
google-auth>=2.34.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
gspread>=6.0.0
opencv-python>=4.10.0.84
numpy>=1.26.0
pytesseract>=0.3.10
requests>=2.31.0
python-dotenv>=1.0.1

